requirements.txt
streamlit
chromadb
sentence-transformers
requests
anthropic
python-dotenv
pytest
beautifulsoup4
